requirements.txt
alembic==1.13.1
asyncpg==0.29.0
boto3==1.34.79
email-validator==2.1.1
fastapi==0.110.0
geoalchemy2==0.14.5
greenlet==3.0.3
httpx==0.26.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
psycopg[binary]==3.1.18
pydantic==2.6.4
pydantic-settings==2.2.1
PyJWT[crypto]==2.8.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-socketio[asyncio_client]==5.11.0
redis==5.0.3
structlog==23.3.0
shapely==2.0.3
stripe==8.7.0
cryptography==42.0.5
firebase-admin==6.4.0
uvicorn[standard]==0.27.1
